change directory structure and add src into sprocket.bin
