+++ example/vcc2018/conf/default/speaker_default_16000.yml
