+++ example/vcc2018/conf/default/speaker_default_22050.yml
