+++ example/vcc2018/conf/default/speaker_default_48000.yml
