+++ example/vcc2018/initialize.py
