+++ sprocket/util/__init__.py
@@ -3,4 +3,3 @@ from .distance import melcd
 from .extfrm import extfrm
 from .hdf5 import HDF5
 from .twf import estimate_twf, align_data
-from .filter import low_pass_filter, high_pass_filter
