+++ sprocket/util/misc.py
@@ -2,38 +2,8 @@
 
 import os
 import numpy as np
-from scipy.signal import firwin, lfilter
 
-from sprocket.util import HDF5, extfrm, static_delta
-
-
-def low_cut_filter(x, fs, cutoff=70):
-    """Low cut filter
-
-    Parameters
-    ---------
-    x : array, shape(`samples`)
-        Waveform sequence
-    fs: array, int
-        Sampling frequency
-    cutoff : float, optional
-        Cutoff frequency of low cut filter
-        Default set to 70 [Hz]
-
-    Returns
-    ---------
-    lcf_x : array, shape(`samples`)
-        Low cut filtered waveform sequence
-    """
-
-    nyquist = fs // 2
-    norm_cutoff = cutoff / nyquist
-
-    # low cut filter
-    fil = firwin(255, norm_cutoff, pass_zero=False)
-    lcf_x = lfilter(fil, 1, x)
-
-    return lcf_x
+from . import HDF5, extfrm, static_delta
 
 
 def read_feats(listf, h5dir, ext='mcep'):
