+++ sprocket/util/yml.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import os
-
 import yaml
